backend/requirements.txt
# Core API - pinned to work with tensorflow-cpu 2.13.0 typing-extensions<4.6.0
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Database - pinned for typing-extensions compatibility
sqlalchemy==1.4.49
psycopg2-binary==2.9.9
alembic==1.11.3
aiosqlite==0.19.0

# ML - PINNED FOR COMPATIBILITY (DO NOT CHANGE)
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
tensorflow-cpu==2.13.0
Pillow==10.1.0
joblib==1.3.2

# typing-extensions pinned to satisfy tensorflow-cpu<4.6.0 AND fastapi>=4.2.0
typing-extensions==4.5.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pydantic==1.10.13